redef Queue, Stack as wrappers of collections.deque
