+++ projects/graph/graph.py
@@ -1,7 +1,14 @@
-"""
-Simple graph implementation
-"""
-from .util import Stack, Queue    # These may come in handy
+############################################################
+#   GRAPH
+#-----------------------------------------------------------
+#   A simple implementation.
+############################################################
+
+from tools.data_structures import Stack, Queue
+
+############################################################
+#   Graph
+############################################################
 
 
 class Graph:
