+++ tools/data_structures.py
@@ -1,0 +1,64 @@
+############################################################
+#   DATA STRUCTURES
+############################################################
+
+from collections import deque
+
+############################################################
+#   Queue
+############################################################
+
+
+class Queue:
+
+    def __init__(self):
+
+        self.__container = deque()
+        return
+
+    def __len__(self):
+
+        return len(self.__container)
+
+    def push(self, value):
+
+        self.__container.append(value)
+        return
+
+    def pop(self):
+
+        if len(self) > 0:
+            return self.__container.popleft()
+
+        else:
+            return None
+
+
+############################################################
+#   Stack
+############################################################
+
+
+class Stack:
+
+    def __init__(self):
+
+        self.__container = deque()
+        return
+
+    def __len__(self):
+
+        return len(self.__container)
+
+    def push(self, value):
+
+        self.__container.appendleft(value)
+        return
+
+    def pop(self):
+
+        if len(self) > 0:
+            return self.__container.popleft()
+
+        else:
+            return None
